LayoutTests/fast/mediastream/peerconnection-Attributes-expected.txt

Tests DeprecatedPeerConnection related Attributes according to http://www.w3.org/TR/webrtc/

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof pc.addStream == 'function' is true
PASS typeof pc.removeStream == 'function' is true
PASS typeof pc.close == 'function' is true
PASS pc.NEW ==0 is true
PASS pc.NEGOTIATING == 1 is true
PASS pc.ACTIVE == 2 is true
PASS pc.CLOSED == 3 is true
FAIL pc.ICE_GATHERING == 0x100  should be true. Was false.
FAIL pc.ICE_WAITING == 0x200 should be true. Was false.
FAIL pc.ICE_CHECKING == 0x300 should be true. Was false.
FAIL pc.ICE_CONNECTED == 0x400 should be true. Was false.
FAIL pc.ICE_COMPLETED == 0x500 should be true. Was false.
FAIL pc.ICE_FAILED == 0x600 should be true. Was false.
FAIL pc.ICE_CLOSED == 0x700 should be true. Was false.
FAIL pc.SDP_IDLE == 0x1000 should be true. Was false.
FAIL pc.SDP_WAITING == 0x2000 should be true. Was false.
FAIL pc.SDP_GLARE ==0x3000 should be true. Was false.
PASS typeof pc.localStreams == 'object' is true
PASS typeof pc.remoteStreams == 'object' is true
PASS typeof pc.onaddstream == 'object' is true
PASS typeof pc.onremovestream == 'object' is true
PASS typeof pc.onconnecting == 'object' is true
PASS typeof pc.onopen == 'object' is true
PASS successfullyParsed is true

TEST COMPLETE
